checker: `s[i]` yields `u8`, the byte type, not `i32` (#5629)
Acceptance item 2 of the char/byte split. A byte and "some integer" were
the same type, so byte-level code type-checked against arbitrary
arithmetic and the byte/scalar distinction survived only as a naming
convention. `char` was already distinct; `u8` was not.

The type-side change is two lines -- both `*ast.Index` string branches
(StringType and the `str` view) now return `NumberType{Width: 8}`. The
rest is the migration it forces, which is atomic: the tree does not
compile in between.

What the migration actually required, beyond the casts:

* The 21 byte-receiver methods in std/i32.fern move to `u8` receivers
  (`is_ascii_digit`, `to_ascii_lower`, `digit_value`, ...), with the two
  case folds returning `u8` as well -- a byte fold yields a byte. Without
  this every `s[i].is_ascii_digit()` orphans: no `as i32` fixes a method
  call without `(s[i] as i32).to_ascii_lower()`, which defeats the point.
  Slices 3 and 4 of the issue plan are mutually dependent, not sequential.
* Six byte-taking helpers flip at their DEFINITIONS rather than their call
  sites -- `__is_ascii_ws`, `__utf8_is_cont`, `__byte_in_set`,
  `__radix_digit`, `_dig`, `repeat_char` -- which removes ~50 argument
  mismatches in one edit.
* `u8` gains Display/Eq/Ord/Hash in core/cmp. Display needs a real body,
  not the empty adopt-`to_string` form the other scalars use: `u8` has no
  scalar module of its own, so nothing supplies `to_string`.

Fern has no implicit unsigned widening, so a byte reaching i32 arithmetic
takes an explicit `as i32`. The cast must land BEFORE any shift:
`(s[i] & 63) << 6` computed in u8 overflows 8 bits and would silently
corrupt multi-byte UTF-8 decoding. The checker catches it as E009 rather
than miscompiling, but the decoders in std/string and std/utf8 are
rewritten to widen first.

No representation change -- `u8` occupies the same slot as `i32`.

Tests: three checker tests replace the one that pinned `s[i]` as i32
(u8 return, i32 rejected, `as i32` accepted); a new e2e case set pins the
runtime half on x86-64 + arm64, including the shift-precedence trap and
2-/3-byte UTF-8 decodes; the interp and IR tests that returned a bare
`s[1]` from an i32 function now widen explicitly.

Verified: all 59 stdlib modules check clean through the import path; 154
example test programs produce byte-identical output against the
pre-change compiler (the 3 excluded are timing benches, nondeterministic
run-to-run); checker, parser, stdlib, interp, ir, modload and literate
suites pass.

+++ cmd/unicodegen/main.go
@@ -1504,7 +1504,7 @@ function _ascii_fold(s: string, from: i32, to: i32): string {
     var buf: u8[] = __alloc_u8(n);
     var i: i32 = 0;
     while (i < n) {
-        var b: i32 = s[i];
+        var b: i32 = s[i] as i32;
         if (b >= from && b < from + 26) { b = b + (to - from); }
         buf = buf.with(i, b as u8);
         i = i + 1;
@@ -1720,8 +1720,8 @@ pub function eq_ignore_case(a: string, b: string): boolean {
         if (na != b.len()) { return false; }
         var k: i32 = 0;
         while (k < na) {
-            var ca: i32 = a[k];
-            var cb: i32 = b[k];
+            var ca: i32 = a[k] as i32;
+            var cb: i32 = b[k] as i32;
             if (ca >= 65 && ca <= 90) { ca = ca + 32; }
             if (cb >= 65 && cb <= 90) { cb = cb + 32; }
             if (ca != cb) { return false; }
@@ -1750,7 +1750,7 @@ pub function swap_case(s: string): string {
         var buf: u8[] = __alloc_u8(n);
         var k: i32 = 0;
         while (k < n) {
-            var b: i32 = s[k];
+            var b: i32 = s[k] as i32;
             if (b >= 65 && b <= 90) { b = b + 32; }
             else if (b >= 97 && b <= 122) { b = b - 32; }
             buf = buf.with(k, b as u8);
