+++ internal/e2eselfhost/self_host_asm_ir_path_test.go
@@ -1405,7 +1405,7 @@ func TestSelfHostAsmIRPath(t *testing.T) {
 		// with asm.fern's register-ABI __fern_chr. Covers len, indexed byte, and
 		// `+` concat of two chr results.
 		{"chr-len", `function main(): i32 { return chr(65).len(); }`},
-		{"chr-byte", `function main(): i32 { return chr(122)[0]; }`},
+		{"chr-byte", `function main(): i32 { return chr(122)[0] as i32; }`},
 		{"chr-concat", `function main(): i32 { var s = chr(72) + chr(105); return s.len() * 100 + s[0]; }`},
 		// String chars -> string[] of 1-char strings (op_str_chars; result is_arr +
 		// is_strarr like split). AST emits __fern_str_chars; IR emits emit_ir_str_chars.
@@ -1791,7 +1791,7 @@ func TestSelfHostAsmIRPath(t *testing.T) {
 		// Two random_i32 draws differ (a stuck/zero generator returns 0/1).
 		{"random-i32-varies", `function main(): i32 { var a: i32 = random_i32(); var b: i32 = random_i32(); if (a == 0) { return 0; } if (a == b) { return 1; } return 7; }`, 7},
 		// A random byte is in 0..255.
-		{"random-bytes-byte-range", `function main(): i32 { var s: string = random_bytes(4); var x: i32 = s[0]; if (x >= 0) { if (x <= 255) { return 1; } } return 0; }`, 1},
+		{"random-bytes-byte-range", `function main(): i32 { var s: string = random_bytes(4); var x: i32 = s[0] as i32; if (x >= 0) { if (x <= 255) { return 1; } } return 0; }`, 1},
 		// uuid_v4: 36 chars, '4' at index 14, '-' at 8, distinct draws.
 		{"uuid-v4", uuidV4Program, 0},
 		// Range-for through the x86-64 self-host IR path (#2699). The legacy
