+++ internal/e2eselfhost/self_host_ir_string_field_rc_test.go
@@ -47,7 +47,7 @@ function rebuild(src: Node[]): Node[] {
 	var out: Node[] = [];
 	var i: i32 = 0;
 	while (i < src.len()) {
-		var s: Node = src[i];
+		var s: Node = src[i] as i32;
 		out = out.append(Node { name: s.name, kids: s.kids, n: s.n });
 		i = i + 1;
 	}
