+++ internal/e2eselfhost/self_host_url_query_ir_test.go
@@ -37,7 +37,7 @@ function url_decode(s: string): string {
     var out: string = "";
     var i: i32 = 0;
     while (i < n) {
-        var b: i32 = s[i];
+        var b: i32 = s[i] as i32;
         var emit: string = s[i:i+1];
         var consumed: i32 = 1;
         if (b == 37 && i + 2 < n) {
